Reorganize code for registry fix
