+++ lib/base.js
@@ -1,113 +1,10 @@
 'use strict';
 
+var _ = require('lodash');
+
 var uptown = require('uptown');
 var createClass = uptown.createClass;
-
-var BaseElement;
-
-/*
- * A refract element implementation with an extensible element registry.
- *
- * The element registry allows you to register your own classes to be instantiated
- * when a particular refract element is encountered, and allows you to specify
- * which elements get instantiated for existing Javascript objects.
- */
-
-var ElementRegistry = createClass({
-  constructor: function() {
-    this.elementMap = {};
-    this.elementDetection = [];
-  },
-
-  /*
-   * Register a new element class for an element.
-   */
-  register: function(name, ElementClass) {
-    this.elementMap[name] = ElementClass;
-    return this;
-  },
-
-  /*
-   * Unregister a previously registered class for an element.
-   */
-  unregister: function(name) {
-    delete this.elementMap[name];
-    return this;
-  },
-
-  /*
-   * Add a new detection function to determine which element
-   * class to use when converting existing js instances into
-   * refract element.
-   */
-  detect: function(test, ElementClass, givenPrepend) {
-    var prepend = givenPrepend === undefined ? true : givenPrepend;
-
-    if (prepend) {
-      this.elementDetection.unshift([test, ElementClass]);
-    } else {
-      this.elementDetection.push([test, ElementClass]);
-    }
-
-    return this;
-  },
-
-  /*
-   * Convert an existing Javascript object into refract element instances, which
-   * can be further processed or serialized into refract or compact refract.
-   * If the item passed in is already refracted, then it is returned
-   * unmodified.
-   */
-  toElement: function(value) {
-    if (value instanceof BaseElement) { return value; }
-
-    var element;
-
-    for(var i = 0; i < this.elementDetection.length; i++) {
-      var test = this.elementDetection[i][0];
-      var ElementClass = this.elementDetection[i][1];
-
-      if (test(value)) {
-        element = new ElementClass(value);
-        break;
-      }
-    }
-
-    return element;
-  },
-
-  /*
-   * Get an element class given an element name.
-   */
-  getElementClass: function(element) {
-    var ElementClass = this.elementMap[element];
-
-    if (ElementClass === undefined) {
-      // Fall back to the base element. We may not know what
-      // to do with the `content`, but downstream software
-      // may know.
-      return BaseElement;
-    }
-
-    return ElementClass;
-  },
-
-  /*
-   * Convert a long-form refract document into refract element instances.
-   */
-   fromRefract: function(doc) {
-    var ElementClass = this.getElementClass(doc.element);
-    return new ElementClass().fromRefract(doc);
-  },
-
-  /*
-   * Convert a compact refract tuple into refract element instances.
-   */
-  fromCompactRefract: function(tuple) {
-    var ElementClass = this.getElementClass(tuple[0]);
-    return new ElementClass().fromCompactRefract(tuple);
-  }
-});
+var ElementRegistry = require('./registry').ElementRegistry;
 
 // Initiate a default Minim registry
 var registry = new ElementRegistry();
@@ -115,7 +12,10 @@ var registry = new ElementRegistry();
 // Use dependency injection so we can split up into multiple files yet resolve
 // any circuluar dependencies. BaseElement is also defined above in order to
 // satisfy any linting issues.
-BaseElement = require('./primitives/base-element')(registry);
+var BaseElement = require('./primitives/base-element')(registry);
+
+registry.BaseElement = BaseElement;
+
 var NullElement = require('./primitives/null-element')(BaseElement);
 var StringElement = require('./primitives/string-element')(BaseElement);
 var NumberElement = require('./primitives/number-element')(BaseElement);
@@ -123,6 +23,26 @@ var BooleanElement = require('./primitives/boolean-element')(BaseElement);
 var ArrayElement = require('./primitives/array-element')(BaseElement, registry);
 var MemberElement = require('./primitives/member-element')(BaseElement, registry);
 var ObjectElement = require('./primitives/object-element')(BaseElement, ArrayElement, MemberElement);
+
+// Set up classes for default elements
+registry
+  .register('null', NullElement)
+  .register('string', StringElement)
+  .register('number', NumberElement)
+  .register('boolean', BooleanElement)
+  .register('array', ArrayElement)
+  .register('object', ObjectElement)
+  .register('member', MemberElement);
+
+// Add instance detection functions to convert existing objects into
+// the corresponding refract elements.
+registry
+  .detect(_.isNull, NullElement, false)
+  .detect(_.isString, StringElement, false)
+  .detect(_.isNumber, NumberElement, false)
+  .detect(_.isBoolean, BooleanElement, false)
+  .detect(_.isArray, ArrayElement, false)
+  .detect(_.isObject, ObjectElement, false);
 
 module.exports = {
   BaseElement: BaseElement,
